edits, move TimeseriesLineChart into own folder
