+++ packages/ui-components/src/components/ShareButton/CopyLinkButton.tsx
@@ -3,13 +3,12 @@ import { CopyToClipboard } from "react-copy-to-clipboard";
 import LinkIcon from "@mui/icons-material/Link";
 import Button from "@mui/material/Button";
 
-export const CopyLinkButton = ({
-  url,
-  onClick,
-}: {
+interface CopyLinkButtonProps {
   url: string;
   onClick: () => void;
-}) => {
+}
+
+export const CopyLinkButton = ({ url, onClick }: CopyLinkButtonProps) => {
   const [copiedLink, setCopiedLink] = useState(false);
   return (
     <CopyToClipboard
