+++ packages/ui-components/src/components/TimeseriesLineChart/index.ts
@@ -1,0 +1,2 @@
+export * from "./interfaces";
+export * from "./TimeseriesLineChart";
